✨ feat: temporal implementations
